pyproject.toml
[project]
name = "tractian-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyyaml>=6.0",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
]
